Change the way Client works in. Introduce session for long running emitters.
Previous emitter's solution was quite fine, but it had three design flaws:

1. Manual control on sending Header/Terminator records from emitter - that are
service records with fixed position in protocol and you shouldn't care about
them.

2. No way to make long running emitter since it was able to send data within
single ASTM communication session.

3. Weird start() method and others.

These problems are fixed now and you're able to use emitter in way it was
assumed to. For example, this is how it could be used with couchdb-python:

def emitter(session):
  header = Header(password='secret')
  for row in db.changes(feed='continuous', include_docs=True, filter='orders/new'):
    with session(header):
      patient = make_patient_record(row.doc)
      assert (yield patient)
      order = make_order_record(row.doc)
      assert (yield order)